requirements.txt
Flask==3.1.0
tensorflow==2.15.0
opencv-python==4.9.0.80
numpy==1.26.4
pyserial==3.5
pyfingerprint==1.5
